pyproject.toml
[project]
name = "news-recommender-system"
version = "0.1.0"
description = ""
authors = [
    { name = "Gaurav Chaudhari", email = "gauravchaudhari37@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0,<2.0.0",
    "numpy>=2.2.3,<3.0.0",
    "sentence-transformers>=3.4.1,<4.0.0",
    "flask>=3.1.0,<4.0.0"
]
# Specify the package (source) directory:
packages = [
    { include = "news_recommender_system" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
